+++ feed/index.html
@@ -72,16 +72,6 @@
                             <span class="border-bottom border-primary border-5 px-1">Posts</span>
                         </li>
                     </ul>
-
-                    <div>
-                        <div id="filter" class="dropdown">
-                            <select id="order-By" class="form-select" aria-label="sort posts">
-                                <option value="title" selected>Title</option>
-                                <option value="newest">Newest</option>
-                                <option value="oldest">Oldest</option>
-                            </select>
-                        </div>
-                    </div>
                 </div>
 
                 <div id="spinnerPosts" class="gap-1 text-center my-5 py-5">
@@ -116,6 +106,7 @@
                                     <p id="bodyTitle"></p>
                                     <div>
                                         <p id="bodyPost" class="card-text"><span id="viewPost"></span></p>
+                                        <!-- ------------------------------To do: add the tags query--------------------- -->
                                         <!-- <p id="tags"></p> -->
                                     </div>
