+++ src/js/profile/deletePost.mjs
@@ -1,6 +1,7 @@
 import { API_KEY, API_BASE, API_POSTS } from "../settings.mjs";
 import { load } from "../shared/storage.mjs";
 import { ErrorHandler } from "../shared/errorHandler.mjs";
+import { displaySpinner } from "./profile.mjs";
 
 // -------------------------2. types-----------------------------
 
@@ -61,7 +62,7 @@ let data = undefined;
  */
 export async function fetchDeletePost(id) {
 
-    // displaySpinner(true);
+    displaySpinner(true);
 
     try {
         const url = API_BASE + API_POSTS + `/${id}`;
@@ -86,9 +87,6 @@ export async function fetchDeletePost(id) {
     } catch (ev) {
         displayError(id, true, "Could not show the post! Please retry later.");
     } finally {
-        // displaySpinner(false);
+        displaySpinner(false);
     }
 }
-
-// fetchDeletePost(id);
-
