+++ src/js/settings.mjs
@@ -8,11 +8,11 @@ export const API_AUTH = "/auth";
 
 export const API_POSTS = "/social/posts";
 
-//https://v2.api.noroff.dev/social/profiles/Fernanda/posts/?_author=true&_comments=true&_reactions=true
+//https://v2.api.noroff.dev/social/profiles/Fernanda/posts/?_author=true&_comments=true&_reactions=true 
 export const API_POSTS_PROFILE = (name) => `/social/profiles/${name}/posts/?_author=true&_comments=true&_reactions=true`;
 
-// export const API_GET_POSTS_PARAMS = "?_author=true&limit=10";
-// export const API_GET_POSTS_PARAMS = "?_author=true&_comments=true&_reactions=true";
+// export const API_GET_POSTS_PARAMS = "?_author=true&limit=10"; // used for reload page API quickly in the development
+// export const API_GET_POSTS_PARAMS = "?_author=true&_comments=true&_reactions=true"; // To do
 export const API_GET_POSTS_PARAMS = "?_author=true";
 
 export const API_SEARCH = "/social/posts/search?_author=true&q=";
